+++ core/switch_core/db/stores/__init__.py
@@ -6,6 +6,7 @@ from switch_core.db.stores.collaboration_bridge_store import CollaborationBridge
 from switch_core.db.stores.document_store import DocumentStore
 from switch_core.db.stores.external_user_store import ExternalUserStore
 from switch_core.db.stores.reference_store import ReferenceStore
+from switch_core.db.stores.reference_type_store import ReferenceTypeStore
 from switch_core.db.stores.room_group_store import RoomGroupStore
 from switch_core.db.stores.room_link_store import RoomLinkStore
 from switch_core.db.stores.room_store import RoomStore
@@ -22,6 +23,7 @@ __all__ = [
     "DocumentStore",
     "ExternalUserStore",
     "ReferenceStore",
+    "ReferenceTypeStore",
     "RoomGroupStore",
     "RoomLinkStore",
     "RoomStore",